requirements.txt
pyTelegramBotAPI==4.23.0
python-dotenv==1.0.1
schedule==1.2.2
tinydb==4.8.0
